+++ docs/docs/changelog.md
@@ -2,8 +2,10 @@
 
 ### 2.2.1
 
-#### 新特效
+#### 新特性
 
+- DatetimePicker
+  - 新增 `default-value` 属性，支持设置面板默认展示值 (by [@yawuling](https://github.com/yawuling) ) )
 - Input
   - 添加自定义外部类名 `custom-textarea-container-class` (by [@yawuling](https://github.com/yawuling) ) )
 - Upload
@@ -11,6 +13,9 @@
 
 #### 优化
 
+- PickerView
+  - 优化多次触发 `change` 事件的问题 (by [@yawuling](https://github.com/yawuling) ) )
+  - 优化多列数据的情况下 `value` 为空时，默认触发一次选中首个选项 (by [@yawuling](https://github.com/yawuling) ) )
 - StatusTip
   - 占位图改用网络地址，小程序打包时会将未使用的组件也打包进去，如果使用本地图片，会明显增大小程序包体积 (by [@yawuling](https://github.com/yawuling) ) )